Изменил класс

+++ rcnn/preprocess_rcnn_utils.py
@@ -2,6 +2,7 @@ import xml.etree.ElementTree as ET
 from os import listdir
 import random
 
+import numpy as np
 import torch
 from torch.utils.data import Dataset
 import cv2
@@ -37,7 +38,7 @@ class RCnnDataset(Dataset):
 
     def _load_image(self, img_path: str):
         image = cv2.imread(img_path, cv2.IMREAD_COLOR)
-        image = cv2.cvtColor(image, cv2.COLOR_BGR2RGB)
+        image = cv2.cvtColor(image, cv2.COLOR_BGR2RGB).astype(np.float32)
         image = image / 255.0
         return image
